requirements.txt
aiohttp==3.8.1
aiohttp-apispec==2.2.3
aiohttp-session==2.11.0
alembic==1.9.4
asyncpg==0.27.0
black==22.6.0
cffi==1.15.1
cryptography==39.0.2
marshmallow-dataclass==8.5.11
pytest==7.1.2
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
pytest-env==0.6.2
pytest-mock==3.10.0
PyYAML==6.0
SQLAlchemy==2.0.4
